pytheas-core shouldn't depend on slf4j binding
Currently pytheas-core has a direct runtime dependency on slf4j-simple
which often results in multiple slf4j bindings in the classpath
of the actual app. The pytheas-core lib should only depend on
the api.

This change also bumps the guava and slf4j versions to match the
common versions used internally right now.